adding jquery styling to user portfolio page

+++ db/migrate/20140924174141_create_investments.rb
@@ -3,7 +3,7 @@ class CreateInvestments < ActiveRecord::Migration
     create_table :investments do |t|
       t.belongs_to :user
       t.belongs_to :stock
-      t.float :number_of_shares
+      t.integer :number_of_shares
       t.float :value
 
       t.timestamps
